app/app_requirements.txt
mlflow
Flask
Pillow
torch
torchvision
pyyaml
gunicorn
eventlet